@@ -13,7 +13,7 @@ import com.rejak.springsecurityexamples.user.usecase.UsersUseCaseImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.*;
 
 @RestController
